edges internal to nodes
Though not perfect, and probably not the right approach, the edges are
now internal properties of the nodes. The difficulty comes in
eliminating or preventing duplicate edges between nodes and
simultaneously computing forces for both nodes. With an internal list
of connected nodes, it’s easy to calculate forces, though rendering
duplicate edges seems a waste of CPU. Also, if at some point it were
desirable to specify connections arbitrarily between nodes, perhaps it
best if they were kept external to the Node class.