minor updates
changed where the Image object loadPixels() and updatePixels() methods
are called when performing the sort. not sure if this will slow things
down or if having it elsewhere causes other problems.

noticed that the evaluation logic was preventing ants from actually
going straight in some cases.

also am playing around with the order of applying the orientation
changes to the ant before swapping the pixels. Not sure if this has
much effect, if any.