Not Much Different from Vector Drift
Able to generate a grid to display the hue and brightness as vectors
and use those vectors as force vectors for positioning nodes.

Each node has an PImage object (need to decide whether this stays as an
PImage or to use color[].

The idea is that the previous frame will be copied by the nodes and
then the image fragments captured by the nodes will be pasted once
their position has been update by applying any force vectors.

Would love to have mouse interactions, rotational velocity, and other
force interactions coded in.